Replaced all the exception throwing ín the Db_Query*::__toString() methods with error sending through the generated SQL.
Exceptions cannot be thrown from a __toString() method when called in string context, in PHP 5.2 and greater.

This change makes it possible to save the error message in the generated SQL itself and then later retrieve it to throw an exception.
The downside is that evaluation of the sql data tree won't be interrupted as it does when throwing an exception.