Added a possibility to easily prefix the first and suffix the last condition in the WHERE part of the MapperSelect queries
This is to make it easier to add conditions that encompass all the conditions the user adds.

Eg. A user wants all the posts of a certain user, this WHERE is generated:
WHERE user_id = 3

Then the user adds his own conditions, but he adds an OR:
WHERE user_id = 3 AND title LIKE '%foo%' OR content LIKE '%bar%'
Then the condition is ruined, because user_id = 3 is not validated across all the rows, only the ones matching title.

To fix this, the user_id = 3 is added as a prefix with an AND and then also an ending parenthesis:
WHERE user_id = 3 AND (title LIKE '%foo%' OR content LIKE '%bar%')

prefix: "user_id = 3 AND(", suffix: ")"